+++ tests/meteora.rs
@@ -2,5 +2,5 @@
 mod meteora_instructions;
 #[path = "meteora/meteora_daam.rs"]
 mod meteora_meteora_daam;
-#[path = "meteora/meteora_dllm.rs"]
-mod meteora_meteora_dllm;
+#[path = "meteora/meteora_dlmm.rs"]
+mod meteora_meteora_dlmm;
