+++ tests/meteora/meteora_dlmm.rs
@@ -2,14 +2,14 @@
 #![allow(deprecated)]
 mod tests {
     use substreams::hex;
-    use substreams_solana_idls::meteora::dllm;
+    use substreams_solana_idls::meteora::dlmm;
 
     #[test]
-    fn unpack_dllm_swap_event() {
+    fn unpack_dlmm_swap_event() {
         // https://solscan.io/tx/<unknown? maybe but not necessary>. Maybe not.
         let bytes = hex!("e445a52e51cb9a1d516ce3becdd00ac471739c6f45944a0e56b3a0f9b2399f421b0a82b56c5b87acc78023b8ed27cc7cf74ba51f6820a847117a8fad48c80ae298e20bf37ff35e7ae8c7b85272b5010701feffff01feffff1f660a0000000000702c000000000000016037000000000000c40200000000000000623d010000000000000000000000000000000000000000");
-        match dllm::anchor_cpi_event::unpack(&bytes).expect("decode event") {
-            dllm::anchor_cpi_event::MeteoraDllmAnchorCpiEvent::Swap(event) => {
+        match dlmm::anchor_cpi_event::unpack(&bytes).expect("decode event") {
+            dlmm::anchor_cpi_event::MeteoraDlmmAnchorCpiEvent::Swap(event) => {
                 assert_eq!(event.lb_pair.to_string(), "8dsKNwMDMh1Vfr4YevzRY9oDS7N29fRHpPXcoDmUVSBd", "lb_pair");
                 assert_eq!(event.from.to_string(), "HeLbvj7KM5fkduCdiufTa5SmVtuuVjySPps7dnp2pDZG", "from");
                 assert_eq!(event.start_bin_id, -511, "start_bin_id");
